+++ src/Request/Behaviours/GetOrderByFromParameterBag.php
@@ -2,15 +2,10 @@
 
 namespace Somnambulist\Bundles\ApiBundle\Request\Behaviours;
 
+use Somnambulist\Bundles\ApiBundle\Request\Filters\ConvertOrderByToArray;
 use Symfony\Component\HttpFoundation\ParameterBag;
 
-use function array_filter;
-use function explode;
 use function is_null;
-use function mb_strtoupper;
-use function str_contains;
-use function substr;
-use function trim;
 
 trait GetOrderByFromParameterBag
 {
@@ -33,23 +28,6 @@ trait GetOrderByFromParameterBag
 
     protected function convertOrderByStringToArrayValues(string $string): array
     {
-        $fields = [];
-
-        foreach (array_filter(explode(',', $string)) as $field) {
-            $dir = 'asc';
-
-            if (str_contains($field, ':')) {
-                [$field, $dir] = explode(':', $field);
-            }
-
-            if (str_starts_with(trim($field), '-')) {
-                $field = substr(trim($field), 1);
-                $dir = 'desc';
-            }
-
-            $fields[trim($field)] = mb_strtoupper(trim($dir));
-        }
-
-        return $fields;
+        return (new ConvertOrderByToArray())($string);
     }
 }