+++ src/Request/Contracts/HasFields.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Bundles\ApiBundle\Request\Contracts;
+
+interface HasFields
+{
+    public function fields(): array;
+}