+++ src/Request/Contracts/HasIncludes.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Bundles\ApiBundle\Request\Contracts;
+
+interface HasIncludes
+{
+    public function includes(): array;
+}