+++ src/Request/Contracts/HasMarker.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Bundles\ApiBundle\Request\Contracts;
+
+interface HasMarker
+{
+    public function marker(): ?string;
+}