+++ src/Response/Types/CollectionType.php
@@ -4,6 +4,8 @@ namespace Somnambulist\Bundles\ApiBundle\Response\Types;
 
 use League\Fractal\Resource\Collection as FractalCollection;
 use League\Fractal\Resource\ResourceAbstract;
+use Somnambulist\Bundles\ApiBundle\Request\Contracts\HasFields;
+use Somnambulist\Bundles\ApiBundle\Request\Contracts\HasIncludes;
 use Somnambulist\Bundles\ApiBundle\Request\FormRequest;
 use Somnambulist\Components\Collection\Contracts\Collection;
 
@@ -32,11 +34,15 @@ class CollectionType extends AbstractType
 
     public static function fromFormRequest(FormRequest $request, Collection $resource, string $transformer, string $key = 'data', array $meta = []): self
     {
-        $obj = new self($resource, $transformer, $key, $request->includes(), $request->fields(), $meta);
-
-        return $obj;
+        return new self(
+            $resource,
+            $transformer,
+            $key,
+            $request instanceof HasIncludes ? $request->includes() : [],
+            $request instanceof HasFields ? $request->fields() : [],
+            $meta
+        );
     }
-
 
     public function asResource(): ResourceAbstract
     {